fix: route adapt Next.js pattern
