+++ src/app/notifications/page.tsx
@@ -1,141 +1,11 @@
-"use client";
+import NotificationsPageClient from "./page-client";
 
-import { getNotifications, markNotificationsAsRead } from "@/actions/notification.action";
-import { NotificationsSkeleton } from "@/components/NotificationSkeleton";
-import { Avatar, AvatarImage } from "@/components/ui/avatar";
-import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { formatDistanceToNow } from "date-fns";
-import { HeartIcon, MessageCircleIcon, UserPlusIcon, ArrowLeftIcon } from "lucide-react";
-import Link from 'next/link';
-import { Button } from "@/components/ui/button";
-
-import { useEffect, useState } from "react";
-import toast from "react-hot-toast";
-import { useRouter } from "next/navigation";
-
-type Notifications = Awaited<ReturnType<typeof getNotifications>>;
-type Notification = Notifications[number];
-
-const getNotificationIcon = (type: string) => {
-  switch (type) {
-    case "LIKE":
-      return <HeartIcon className="size-4 text-red-500" />;
-    case "COMMENT":
-      return <MessageCircleIcon className="size-4 text-blue-500" />;
-    case "FOLLOW":
-      return <UserPlusIcon className="size-4 text-green-500" />;
-    default:
-      return null;
-  }
-};
-
-function NotificationsPage() {
-  const [notifications, setNotifications] = useState<Notification[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const router = useRouter();
-
-  useEffect(() => {
-    const fetchNotifications = async () => {
-      setIsLoading(true);
-      try {
-        const data = await getNotifications();
-        setNotifications(data);
-
-        const unreadIds = data.filter((n) => !n.read).map((n) => n.id);
-        if (unreadIds.length > 0) await markNotificationsAsRead(unreadIds);
-      } catch (error) {
-        toast.error("Failed to fetch notifications");
-      } finally {
-        setIsLoading(false);
-      }
-    };
-
-    fetchNotifications();
-  }, []);
-
-  if (isLoading) return <NotificationsSkeleton />;
-
+export default function NotificationsPage() {
   return (
-    <div className="space-y-4">
-      <div className="mb-4">
-        <Button variant="ghost" size="sm" onClick={() => router.back()} className="flex items-center gap-2">
-          <ArrowLeftIcon className="size-5" />
-          Back
-        </Button>
+    <div className="grid grid-cols-1 lg:grid-cols-9 gap-6">
+      <div className="lg:col-span-6">
+        <NotificationsPageClient />
       </div>
-      <Card>
-        <CardHeader className="border-b">
-          <div className="flex items-center justify-between">
-            <CardTitle>Notifications</CardTitle>
-            <span className="text-sm text-muted-foreground">
-              {notifications.filter((n) => !n.read).length} unread
-            </span>
-          </div>
-        </CardHeader>
-        <CardContent className="p-0">
-          <ScrollArea className="h-[calc(90vh-12rem)]">
-            {notifications.length === 0 ? (
-              <div className="p-4 text-center text-muted-foreground">No notifications yet</div>
-            ) : (
-              notifications.map((notification) => (
-                <Link key={notification.id} href={`/post/${notification.post?.id}`} legacyBehavior>
-                  <div
-                    className={`flex items-start gap-4 p-4 border-b hover:bg-muted/25 transition-colors cursor-pointer ${!notification.read ? "bg-muted/50" : ""
-                      }`}
-                  >
-                    <Avatar className="mt-1">
-                      <AvatarImage src={notification.creator.image ?? "/avatar.png"} />
-                    </Avatar>
-                    <div className="flex-1 space-y-1">
-                      <div className="flex items-center gap-2">
-                        {getNotificationIcon(notification.type)}
-                        <span>
-                          <span className="font-medium">
-                            {notification.creator.name ?? notification.creator.username}
-                          </span>{" "}
-                          {notification.type === "FOLLOW"
-                            ? "started following you"
-                            : notification.type === "LIKE"
-                              ? "liked your post"
-                              : "commented on your post"}
-                        </span>
-                      </div>
-
-                      {notification.post &&
-                        (notification.type === "LIKE" || notification.type === "COMMENT") && (
-                          <div className="pl-6 space-y-2">
-                            <div className="text-sm text-muted-foreground rounded-md p-2 bg-muted/30 mt-2">
-                              <p>{notification.post.content}</p>
-                              {notification.post.images[0] && (
-                                <img
-                                  src={notification.post.images[0]}
-                                  alt="Post content"
-                                  className="mt-2 rounded-md w-full max-w-[200px] h-auto object-cover"
-                                />
-                              )}
-                            </div>
-
-                            {notification.type === "COMMENT" && notification.comment && (
-                              <div className="text-sm p-2 bg-accent/50 rounded-md">
-                                {notification.comment.content}
-                              </div>
-                            )}
-                          </div>
-                        )}
-
-                      <p className="text-sm text-muted-foreground pl-6">
-                        {formatDistanceToNow(new Date(notification.createdAt), { addSuffix: true })}
-                      </p>
-                    </div>
-                  </div>
-                </Link>
-              ))
-            )}
-          </ScrollArea>
-        </CardContent>
-      </Card>
     </div>
   );
 }
-export default NotificationsPage;
